fix: re-structure file directories
