+++ client/src/app/pages/advisory-details/components/csaf-overview-tab/ImpactChart.tsx
