+++ client/src/app/pages/advisory-details/components/csaf-source-tab/csaf-source.tsx
