+++ client/src/app/pages/advisory-details/components/csaf-vulnerabilities-tab/VulnerabilityCard.tsx
@@ -1,14 +1,10 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import dayjs from "dayjs";
-
 import {
-  Button,
+  capitalize,
   Card,
   CardBody,
-  CardHeader,
-  CardTitle,
   Content,
   DescriptionList,
   DescriptionListDescription,
@@ -18,12 +14,10 @@ import {
   ExpandableSection,
   Flex,
   FlexItem,
-  Label,
   Stack,
   StackItem,
   Title,
 } from "@patternfly/react-core";
-import AngleRightIcon from "@patternfly/react-icons/dist/esm/icons/angle-right-icon";
 import ExternalLinkAltIcon from "@patternfly/react-icons/dist/esm/icons/external-link-alt-icon";
 
 import { Paths } from "@app/Routes";
@@ -34,12 +28,11 @@ import type {
   Vulnerability,
 } from "@app/specs/csaf/csaf-v2.0-schema";
 
-import { CsafCvssDetails } from "./csaf-cvss-details";
-import { CsafRemediations } from "./csaf-remediations";
-
-const expandableSectionStyle = {
-  "--pf-v6-c-expandable-section--m-expanded__toggle-icon--Rotate": "90deg",
-} as React.CSSProperties;
+import { ExtendedSeverity } from "@app/api/models";
+import { formatDate } from "@app/utils/utils";
+import { CsafRemediations } from "./Remediations";
+import { AffectedProducts } from "./AffectedProducts";
+import { CsafCvssDetails } from "./CvssDetails";
 
 type CvssV3 =
   | JSONSchemaForCommonVulnerabilityScoringSystemVersion30
@@ -49,45 +42,6 @@ interface CsafVulnerabilityCardProps {
   vulnerability: Vulnerability;
   productNameMap: Map<string, string>;
 }
-
-const INITIAL_PRODUCTS_SHOWN = 5;
-
-const AffectedProducts: React.FC<{
-  productIds: string[];
-  productNameMap: Map<string, string>;
-}> = ({ productIds, productNameMap }) => {
-  const [expanded, setExpanded] = React.useState(false);
-  const visible = expanded
-    ? productIds
-    : productIds.slice(0, INITIAL_PRODUCTS_SHOWN);
-  const remaining = productIds.length - INITIAL_PRODUCTS_SHOWN;
-
-  return (
-    <Flex gap={{ default: "gapXs" }} flexWrap={{ default: "wrap" }}>
-      {visible.map((id) => (
-        <FlexItem key={id}>
-          <Label variant="outline" color="orange" isCompact>
-            {productNameMap.get(id) ?? id}
-          </Label>
-        </FlexItem>
-      ))}
-      {!expanded && remaining > 0 && (
-        <FlexItem>
-          <Button variant="link" size="sm" onClick={() => setExpanded(true)}>
-            +{remaining} more
-          </Button>
-        </FlexItem>
-      )}
-      {expanded && productIds.length > INITIAL_PRODUCTS_SHOWN && (
-        <FlexItem>
-          <Button variant="link" size="sm" onClick={() => setExpanded(false)}>
-            Show less
-          </Button>
-        </FlexItem>
-      )}
-    </Flex>
-  );
-};
 
 export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
   vulnerability,
@@ -103,52 +57,33 @@ export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
 
   return (
     <Card>
-      <CardHeader>
-        <CardTitle style={{ fontWeight: 400 }}>
-          <Flex
-            alignItems={{ default: "alignItemsCenter" }}
-            gap={{ default: "gapMd" }}
-          >
-            <FlexItem>
-              {cveLink ? (
-                <Link
-                  to={cveLink}
-                  style={{
-                    fontSize: "var(--pf-v6-global--FontSize--lg)",
-                    textDecoration: "underline",
-                  }}
-                >
-                  {vulnerability.cve}
-                </Link>
-              ) : (
-                <span style={{ fontSize: "var(--pf-v6-global--FontSize--lg)" }}>
-                  {vulnerability.cve}
-                </span>
-              )}
-            </FlexItem>
-            {cvss && (
-              <FlexItem>
-                <SeverityShieldAndText
-                  value={
-                    cvss.baseSeverity.toLowerCase() as
-                      | "critical"
-                      | "high"
-                      | "medium"
-                      | "low"
-                      | "none"
-                  }
-                  score={cvss.baseScore}
-                  showLabel
-                  showScore
-                />
-              </FlexItem>
-            )}
-          </Flex>
-        </CardTitle>
-      </CardHeader>
-
       <CardBody>
         <Stack hasGutter>
+          <StackItem>
+            <Flex
+              alignItems={{ default: "alignItemsCenter" }}
+              gap={{ default: "gapMd" }}
+            >
+              <FlexItem>
+                {cveLink ? (
+                  <Link to={cveLink}>{vulnerability.cve}</Link>
+                ) : (
+                  <span>{vulnerability.cve}</span>
+                )}
+              </FlexItem>
+              {cvss && (
+                <FlexItem>
+                  <SeverityShieldAndText
+                    value={cvss.baseSeverity.toLowerCase() as ExtendedSeverity}
+                    score={cvss.baseScore}
+                    showLabel
+                    showScore
+                  />
+                </FlexItem>
+              )}
+            </Flex>
+          </StackItem>
+
           {vulnerability.title && (
             <StackItem>
               <Content component="p">{vulnerability.title}</Content>
@@ -181,7 +116,7 @@ export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
                 <DescriptionListGroup>
                   <DescriptionListTerm>Discovered</DescriptionListTerm>
                   <DescriptionListDescription>
-                    {dayjs(vulnerability.discovery_date).format("MMM D, YYYY")}
+                    {formatDate(vulnerability.discovery_date)}
                   </DescriptionListDescription>
                 </DescriptionListGroup>
               )}
@@ -189,7 +124,7 @@ export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
                 <DescriptionListGroup>
                   <DescriptionListTerm>Released</DescriptionListTerm>
                   <DescriptionListDescription>
-                    {dayjs(vulnerability.release_date).format("MMM D, YYYY")}
+                    {formatDate(vulnerability.release_date)}
                   </DescriptionListDescription>
                 </DescriptionListGroup>
               )}
@@ -198,9 +133,7 @@ export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
                 vulnerability.threats.map((threat, i) => (
                   <DescriptionListGroup key={`${threat.category}-${i}`}>
                     <DescriptionListTerm>
-                      {threat.category
-                        .replace(/_/g, " ")
-                        .replace(/^\w/, (c) => c.toUpperCase())}
+                      {capitalize(threat.category)}
                     </DescriptionListTerm>
                     <DescriptionListDescription>
                       {threat.details}
@@ -218,18 +151,10 @@ export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
               <StackItem>
                 <ExpandableSection
                   toggleText="CVSS v3 details"
-                  toggleIcon={<AngleRightIcon />}
-                  style={expandableSectionStyle}
                   isExpanded={cvssExpanded}
                   onToggle={(_e, expanded) => setCvssExpanded(expanded)}
                 >
-                  <div
-                    style={{
-                      marginTop: "var(--pf-v6-global--spacer--sm)",
-                    }}
-                  >
-                    <CsafCvssDetails cvss={cvss} />
-                  </div>
+                  <CsafCvssDetails cvss={cvss} />
                 </ExpandableSection>
               </StackItem>
             </>
@@ -243,19 +168,11 @@ export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
               <StackItem>
                 <ExpandableSection
                   toggleText={`Affected products (${affectedProducts.length})`}
-                  toggleIcon={<AngleRightIcon />}
-                  style={expandableSectionStyle}
                 >
-                  <div
-                    style={{
-                      marginTop: "var(--pf-v6-global--spacer--sm)",
-                    }}
-                  >
-                    <AffectedProducts
-                      productIds={affectedProducts}
-                      productNameMap={productNameMap}
-                    />
-                  </div>
+                  <AffectedProducts
+                    productIds={affectedProducts}
+                    productNameMap={productNameMap}
+                  />
                 </ExpandableSection>
               </StackItem>
             </>
